refactor(app): :zap: Package updates; minor bug fixes; revision update to 1.4.6

+++ dms3client/client_config.go
@@ -18,5 +18,5 @@ type structSettings struct {
 type structServer struct {
 	IP            string
 	Port          int
-	CheckInterval uint16
+	CheckInterval int
 }
